[Android] Fixed color for tab groups toolbar
Chromium change:
https://chromium.googlesource.com/chromium/src/+/73c2d3d5ee5865411e0f09909f1456c5e7bbc4e3

Bottom tab strip bg to Surface-3 in night mode.

The previous implementation assumed that the background color of the tab
strip should match the omnibox. This is no longer the case, as the
tab strip should now match infobars/dialogs, and be surface-0 (day) and
surface-3 (night).

So replace the theme provider with an is incognito provider, to let the
view make the appropriate decision.

Bug: 1243098