Removing rapidjson brave fork
New clang version have been triggering warnings on the forked version of
rapidjson being used by brave. There's work under way to remove
rapidjson entirely from the codebase, but that has proven to be a bit
more challanged than initially expected.

This change removes the use of the rapidjson brave fork, which was
created mainly to have a .gn file in the root path, and uses the
original rapidjson repo to clone it from.