[Android] Adjustments to the various upstream java changes
Mostly renaming and removing obsolete code.

Chromium changes:
https://github.com/chromium/chromium/commit/7068c98f7fafe23ea93e203c182f889a04380d2e

Enable All Suggestions to feature Action Chips.
The change renames PedalSuggestionProcessor as ActionChipsProcessor
and relocates it to the `.base` package.

The change also renames OmniboxPedalDelegate as ActionChipsDelegate,
and plumbs the Delegate to all relevant SuggestionProcessors:
- Entity,
- Answer,
- Tail, and
- Basic (catch-all suggestion processor).
to increase the availability of Pedals and Journeys.

Bug: 1418077

https://github.com/chromium/chromium/commit/819d1f2e3fc5b4044abf1b5c0fc24adcab272b04

[Crow] Remove CrowButtonDelegate dependency (Feed+Menu)
Bug: 1408601

https://github.com/chromium/chromium/commit/e82ad92b6f71d5baca1ec94b389c1d7cea903161

Simplify Java TaskTraits
This is essentially prep work for future cleanup, I've minimized the API
exposed from TaskTraits (after 4 years we still haven't needed a lot of
the infrastructure that was built here).

There are no functional changes here, TaskTraits.THREAD_POOL_BEST_EFFORT and TaskTraits.BEST_EFFORT were/are equivalent.

Change-Id: Ida665c009d20d891b25821fec9d603bd9c1a3657

https://github.com/chromium/chromium/commit/2d180440d8b213e2f9b5290dd76998414d0f02ff

[Tab Management] Simplify M5 params
Flatten launch params for M5 to make things a bit easier to reason
about. This CL flattens params for

* ENABLE_LAUNCH_POLISH
* ENABLE_LAUNCH_BUG_FIX
* ENABLE_DEFERRED_FAVICON

into TabUiFeatureUtilities#isTabGroupsAndroidContinuationEnabled(). This
flag has been turned off and will only be re-activated once we are ready
to re-attempt the M5 launch.

This CL also removes

* ENABLE_TAB_GROUP_SHARING

which is deprecated in favor of TabSelectionEditorV2.

The second, third, and fourth bugs are speculatively fixed by this.

Bug: 1420613, 1418922, 1420197, 1419842