Fix a cause of unnecessary create_dist recompiles
It does not seem to be necessary to delete the dist/ output directory
before create_dist: The only script (including devops) that operates
with wildcards in this directory is sign_upload_linux_pkgs. It is
called by CI in a clean repo, it only captures *.deb and *.rpm files
and it has some logic for dealing with existing files.